+++ src/components/@core/navigation-menu/light/right-menu/pages-menu-item.tsx
@@ -1,7 +1,7 @@
 import { MenuDivider, MenuItem, MenuList } from "@chakra-ui/react";
 import LocalLink from "@components/@core/local-link";
 import useGlobalState from "@hooks/use-global-state";
-import { getPagesMenu } from "@utils/pages.util";
+import { getPagesMenu } from "@utils/pages";
 import useTranslation from "next-translate/useTranslation";
 import React, { useMemo } from "react";
